// NextJS/Widgets/RSVP/src/types/widgetData.ts
export interface UserStats {
  Total_Prayers: {
    Label: string;
    Count: number;
  };
  Prayer_Streak: {
    Label: string;
    Count: number;
  };
  Prayers_Today: {
    Label: string;
    Count: number;
  };
}